refactor: 커스텀 훅 관심사 분리
